pyproject.toml
[project]
name = "unplayed"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "spotipy>=2.26.0",
    "sqlite-utils>=3.39",
    "typer>=0.24.1",
]
